refactor(meta): extract table identifier types to ident.rs
- Created table/ident.rs with 5 identifier types
- Moved TableIdent, TableNameIdent, DBIdTableName, TableId, TableIdHistoryIdent
- Kept public API unchanged via pub use ident::*
- Removed unused ToTenant import from mod.rs

Benefits:
- Clear separation of identifier types from table metadata
- Reduced mod.rs from 1614 to 1487 lines
- Better code organization and navigation
